KBV-34 - changes made as per PR comments
