--- src/main/java/uk/gov/di/cri/experian/kbv/api/domain/FieldType.java
@@ -1,6 +1,0 @@
-package uk.gov.di.cri.experian.kbv.api.domain;
-
-public enum FieldType {
-    G,
-    R;
-}
